Fixed a bug where sometimes after restarting the typeracer game with the “Random words” passage type, the game would freeze for several seconds.
Fixed a bug where sometimes after restarting the typeracer game with the “Random words” passage type, the game would freeze for several seconds. This is because every time the game restarted, the game would 1000 random words all at once and add all of the words to the passage, which was extremely slow. The fix was to only generate 100 words at the start of the game, and every time the user finishes typing a word, another word would be generated and added to the passage.